# pkg/api/dialog/dialog.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "dialog/dialog.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "DialogAPI"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/dialog/list": {
      "get": {
        "operationId": "DialogAPI_List",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/postListResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "DialogAPI"
        ]
      }
    },
    "/dialog/{userId}/send": {
      "post": {
        "operationId": "DialogAPI_Send",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/postSendResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "DialogAPI"
        ]
      }
    }
  },
  "definitions": {
    "ListResponseDialogMessage": {
      "type": "object",
      "properties": {
        "from": {
          "type": "string",
          "format": "uint64"
        },
        "to": {
          "type": "string",
          "format": "uint64"
        },
        "text": {
          "type": "string"
        },
        "sentAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "postListResponse": {
      "type": "object",
      "properties": {
        "messages": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/ListResponseDialogMessage"
          }
        }
      }
    },
    "postSendResponse": {
      "type": "object"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}